backend/requirements/base.txt
Django==5.1.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-filter==24.3
django-cors-headers==4.6.0
psycopg2-binary==2.9.10
django-axes==7.0.2
python-dotenv==1.0.1
reportlab==4.2.5
arabic-reshaper==3.0.0
python-bidi==0.4.2
gunicorn==23.0.0
